requirements.txt
celery==5.2.7
Django==4.1.4
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-cron==0.6.0
django-crontab==0.7.1
django-phonenumber-field==7.0.1
django-redis==5.2.0
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
phonenumbers==8.13.2
Pillow==9.3.0
django-environ==0.9.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
redis==4.4.0
gunicorn==20.1.0
python-dateutil==2.8.2
opencv-python==4.7.0.68
pytesseract==0.3.10
django-filter==22.1 